Time: 0 ms (100.00%) | Memory: 7.8 MB (91.36%) - LeetSync

+++ 2649-count-total-number-of-colored-cells/count-total-number-of-colored-cells.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long coloredCells(int n) {
-        long long total = 0; 
+     long long total = 0; 
         int a = 1;          
         int d = 4;          
 
@@ -9,7 +9,7 @@ public:
             return 1; 
         } else if (n > 1) {
             
-            total = (long long)((n / 2.0) * (2 * a + (n - 1) * d)) - (n - 1);
+            total = ((n / 2.0) * (2 * a + (n - 1) * d)) - (n - 1);
         }
 
         return total;
